speedtest-python-back/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
speedtest-cli==2.1.3
tqdm==4.66.2
Werkzeug==3.0.2
zipp==3.18.1
